// src/router/types.ts
// @ts-ignore
import { RouteMeta, RouteRecordRaw } from 'vue-router';
import { Component } from 'vue';

// @ts-ignore
export interface AppRouteRecordRaw extends Omit<RouteRecordRaw, 'meta'> {
  name: string;
  meta: RouteMeta;
  component?: Component | string;
  components?: Component;
  children?: AppRouteRecordRaw[];
  props?: Recordable;
  fullPath?: string;
}